Refactor types out of Covenant.Type, cleanup
