+++ src/Covenant/Type.hs
@@ -36,7 +36,6 @@ where
 
 import Control.Monad (foldM, unless)
 import Control.Monad.Except (ExceptT, MonadError (throwError), catchError, runExceptT)
-import Control.Monad.Reader (MonadReader, Reader, asks, local, runReader)
 import Control.Monad.State.Strict (State, evalState, gets, modify)
 import Covenant.DeBruijn (DeBruijn, asInt)
 import Covenant.Index
@@ -49,13 +48,27 @@ import Covenant.Index
     intCount,
     intIndex,
   )
+import Covenant.Internal.Type
+  ( AbstractTy (BoundAt),
+    BuiltinFlatT
+      ( BLS12_381_G1_ElementT,
+        BLS12_381_G2_ElementT,
+        BLS12_381_MlResultT,
+        BoolT,
+        ByteStringT,
+        IntegerT,
+        StringT,
+        UnitT
+      ),
+    CompT (CompT),
+    Renamed (Rigid, Unifiable, Wildcard),
+    ValT (Abstraction, BuiltinFlat, ThunkT),
+  )
 import Data.Coerce (coerce)
 #if __GLASGOW_HASKELL__==908
 import Data.Foldable (foldl')
 #endif
-import Data.Functor.Classes (Eq1 (liftEq))
 import Data.Kind (Type)
-import Data.List.NonEmpty (NonEmpty ((:|)))
 import Data.Map.Merge.Strict qualified as Merge
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
@@ -69,99 +82,19 @@ import Data.Vector qualified as Vector
 import Data.Vector.NonEmpty (NonEmptyVector)
 import Data.Vector.NonEmpty qualified as NonEmpty
 import Data.Word (Word64)
-import Optics.AffineFold (preview)
-import Optics.At (ix)
-import Optics.Getter (to, view)
-import Optics.Label (LabelOptic (labelOptic))
-import Optics.Lens (A_Lens, Lens', lens)
-import Optics.Optic ((%))
-import Optics.Review (review)
-import Optics.Setter (over, set)
-import Prettyprinter
-  ( Doc,
-    Pretty (pretty),
-    hsep,
-    parens,
-    viaShow,
-    (<+>),
+import Optics.Core
+  ( A_Lens,
+    LabelOptic (labelOptic),
+    ix,
+    lens,
+    over,
+    preview,
+    review,
+    set,
+    to,
+    view,
+    (%),
   )
-
--- | A type abstraction, using a combination of a DeBruijn index (to indicate
--- which scope it refers to) and a positional index (to indicate which bound
--- variable in that scope it refers to).
---
--- = Important note
---
--- This is a /relative/ representation: any given 'AbstractTy' could refer to
--- different things when placed in different positions in the ASG. This stems
--- from how DeBruijn indices behave: 'Z' refers to \'our immediate enclosing
--- scope\', @'S' 'Z'@ to \'one scope outside our immediate enclosing scope\',
--- etc. This can mean different things depending on what these scope(s) are.
---
--- @since 1.0.0
-data AbstractTy = BoundAt DeBruijn (Index "tyvar")
-  deriving stock
-    ( -- | @since 1.0.0
-      Eq,
-      -- | @since 1.0.0
-      Show
-    )
-
--- | @since 1.0.0
-data Renamed
-  = -- | Set by an enclosing scope, and thus is essentially a
-    -- concrete type, we just don't know which. First field is its \'true
-    -- level\', second field is the positional index in that scope.
-    Rigid Int (Index "tyvar")
-  | -- | Can be unified with something, but must be consistent: that is, only one
-    -- unification for every instance. Field is this variable's positional index;
-    -- we don't need to track the scope, as only one scope contains unifiable
-    -- bindings.
-    Unifiable (Index "tyvar")
-  | -- | /Must/ unify with everything, except with other distinct wildcards in the
-    -- same scope. First field is a unique /scope/ identifier, second is the
-    -- positional index within that scope. We must have unique identifiers for
-    -- wildcard scopes, as wildcards unify with everything /except/ other
-    -- wildcards in the same scope.
-    Wildcard Word64 (Index "tyvar")
-  deriving stock
-    ( -- | @since 1.0.0
-      Eq,
-      -- | @since 1.0.0
-      Ord,
-      -- | @since 1.0.0
-      Show
-    )
-
--- | A computation type, with abstractions indicated by the type argument. In
--- pretty much any case imaginable, this would be either 'AbstractTy' (in the
--- ASG), or 'Renamed' (after renaming).
---
--- = Important note
---
--- This type has a \'type abstraction boundary\' just before it: the first field
--- indicates how many type variables it binds.
---
--- The /last/ entry in the 'NonEmpty' indicates the return type.
---
--- @since 1.0.0
-data CompT (a :: Type) = CompT (Count "tyvar") (NonEmptyVector (ValT a))
-  deriving stock
-    ( -- | @since 1.0.0
-      Eq,
-      -- | @since 1.0.0
-      Show
-    )
-
--- Note (Koz, 04/03/2025): We use this for testing to compare for structural
--- similarity.
-instance Eq1 CompT where
-  {-# INLINEABLE liftEq #-}
-  liftEq f (CompT abses1 xs) (CompT abses2 ys) =
-    abses1 == abses2 && liftEq (liftEq f) xs ys
-
-instance Pretty (CompT Renamed) where
-  pretty = runPrettyM . prettyCompTWithContext
 
 -- | Helper for defining the \'bodies\' of computation types, without having to
 -- use 'NonEmptyVector' functions.
@@ -227,40 +160,6 @@ comp2 = CompT count2
 comp3 :: NonEmptyVector (ValT AbstractTy) -> CompT AbstractTy
 comp3 = CompT count3
 
--- | A value type, with abstractions indicated by the type argument. In pretty
--- much any case imaginable, this would be either 'AbstractTy' (in the ASG) or
--- 'Renamed' (after renaming).
---
--- @ since 1.0.0
-data ValT (a :: Type)
-  = -- | An abstract type.
-    Abstraction a
-  | -- | A suspended computation.
-    ThunkT (CompT a)
-  | -- | A builtin type without any nesting.
-    BuiltinFlat BuiltinFlatT
-  deriving stock
-    ( -- | @since 1.0.0
-      Eq,
-      -- | @since 1.0.0
-      Show
-    )
-
--- Note (Koz, 04/03/2025): We use this for testing to compare for structural
--- similarity.
-instance Eq1 ValT where
-  {-# INLINEABLE liftEq #-}
-  liftEq f = \case
-    Abstraction abs1 -> \case
-      Abstraction abs2 -> f abs1 abs2
-      _ -> False
-    ThunkT t1 -> \case
-      ThunkT t2 -> liftEq f t1 t2
-      _ -> False
-    BuiltinFlat t1 -> \case
-      BuiltinFlat t2 -> t1 == t2
-      _ -> False
-
 -- | Helper for defining type variables.
 --
 -- @since 1.0.0
@@ -314,24 +213,6 @@ mlResultT = BuiltinFlat BLS12_381_MlResultT
 -- @since 1.0.0
 unitT :: forall (a :: Type). ValT a
 unitT = BuiltinFlat UnitT
-
--- | All builtin types that are \'flat\': that is, do not have other types
--- \'nested inside them\'.
-data BuiltinFlatT
-  = UnitT
-  | BoolT
-  | IntegerT
-  | StringT
-  | ByteStringT
-  | BLS12_381_G1_ElementT
-  | BLS12_381_G2_ElementT
-  | BLS12_381_MlResultT
-  deriving stock
-    ( -- | @since 1.0.0
-      Eq,
-      -- | @since 1.0.0
-      Show
-    )
 
 -- Used during renaming. Contains a source of fresh indices for wildcards, as
 -- well as tracking:
@@ -721,142 +602,3 @@ substitute index toSub = \case
   ThunkT (CompT abstractions xs) ->
     ThunkT . CompT abstractions . fmap (substitute index toSub) $ xs
   BuiltinFlat t -> BuiltinFlat t
-
--- Pretty printing stuff, somewhat awkward to put it here but doing so avoids orphan instances
-
--- Keeping the field names for clarify even if we don't use them
-newtype ScopeBoundary = ScopeBoundary {_getBoundary :: Int}
-  deriving newtype (Show, Eq, Ord, Num, Enum)
-
-data PrettyContext (ann :: Type)
-  = PrettyContext
-  { _pcBoundIdents :: Map ScopeBoundary [Doc ann],
-    _pcCurrentScope :: ScopeBoundary,
-    _pcVarStream :: [Doc ann]
-  }
-
--- Lazily generated infinite list of variables. Will start with a, b, c... and cycle around to a1, b2, c3 etc.
--- We could do something more sophisticated but this should work.
-infiniteVars :: forall (ann :: Type). [Doc ann]
-infiniteVars =
-  let aToZ = ['a' .. 'z']
-      intStrings = ("" <$ aToZ) <> map (show @Integer) [0 ..]
-   in zipWith (\x xs -> pretty (x : xs)) aToZ intStrings
-
--- optics, for convenience. Manually defined b/c TH splices tend to blow up HLS
-boundIdents :: forall (ann :: Type). Lens' (PrettyContext ann) (Map ScopeBoundary [Doc ann])
-boundIdents = lens goGet goSet
-  where
-    goGet :: PrettyContext ann -> Map ScopeBoundary [Doc ann]
-    goGet (PrettyContext bi _ _) = bi
-
-    goSet :: PrettyContext ann -> Map ScopeBoundary [Doc ann] -> PrettyContext ann
-    goSet (PrettyContext _ scop vars) bi' = PrettyContext bi' scop vars
-
-currentScope :: forall (ann :: Type). Lens' (PrettyContext ann) ScopeBoundary
-currentScope = lens goGet goSet
-  where
-    goGet :: PrettyContext ann -> ScopeBoundary
-    goGet (PrettyContext _ scop _) = scop
-
-    goSet :: PrettyContext ann -> ScopeBoundary -> PrettyContext ann
-    goSet (PrettyContext bi _ vars) scop = PrettyContext bi scop vars
-
-varStream :: forall (ann :: Type). Lens' (PrettyContext ann) [Doc ann]
-varStream = lens goGet goSet
-  where
-    goGet :: PrettyContext ann -> [Doc ann]
-    goGet (PrettyContext _ _ vars) = vars
-
-    goSet :: PrettyContext ann -> [Doc ann] -> PrettyContext ann
-    goSet (PrettyContext bi scop _) = PrettyContext bi scop
-
--- Generate N fresh var names and use the supplied monadic function to do something with them.
-withFreshVarNames :: forall (ann :: Type) (a :: Type). Int -> ([Doc ann] -> PrettyM ann a) -> PrettyM ann a
-withFreshVarNames n act = do
-  stream <- asks (view varStream)
-  let (used, rest) = splitAt n stream
-  local (set varStream rest) $ act used
-
--- Maybe make a newtype with error reporting since this can fail, but do later since *should't* fail
-newtype PrettyM (ann :: Type) (a :: Type) = PrettyM (Reader (PrettyContext ann) a)
-  deriving newtype (Functor, Applicative, Monad, MonadReader (PrettyContext ann))
-
-runPrettyM :: forall (ann :: Type) (a :: Type). PrettyM ann a -> a
-runPrettyM (PrettyM ma) = runReader ma (PrettyContext mempty 0 infiniteVars)
-
-bindVars :: forall (ann :: Type) (a :: Type). Count "tyvar" -> ([Doc ann] -> PrettyM ann a) -> PrettyM ann a
-bindVars count' act
-  | count == 0 = crossBoundary (act [])
-  | otherwise = crossBoundary $ do
-      here <- asks (view currentScope)
-      withFreshVarNames count $ \newBoundVars ->
-        local (over boundIdents (Map.insert here newBoundVars)) (act newBoundVars)
-  where
-    -- Increment the current scope
-    crossBoundary :: PrettyM ann a -> PrettyM ann a
-    crossBoundary = local (over currentScope (+ 1))
-
-    count :: Int
-    count = review intCount count'
-
--- Bad name, but anyway, looks up the Doc for a pretty
-lookupRigid :: forall (ann :: Type). Int -> Index "tyvar" -> PrettyM ann (Doc ann)
-lookupRigid (ScopeBoundary -> scopeOffset) argIndex = do
-  let argIndex' = review intIndex argIndex
-  here <- asks (view currentScope)
-  asks (preview (boundIdents % ix (here + scopeOffset) % ix argIndex')) >>= \case
-    Nothing ->
-      -- TODO: actual error reporting
-      error $
-        "Internal error: The encountered a variable at arg index "
-          <> show argIndex'
-          <> " with true level "
-          <> show scopeOffset
-          <> " but could not locate the corresponding pretty form at scope level "
-          <> show here
-    Just res' -> pure res'
-
-prettyRenamedWithContext :: forall (ann :: Type). Renamed -> PrettyM ann (Doc ann)
-prettyRenamedWithContext = \case
-  Rigid offset index -> lookupRigid offset index
-  Unifiable i -> lookupRigid 0 i -- ok maybe 'lookupRigid' isn't the best name
-  Wildcard w64 i -> pure $ "_" <> viaShow w64 <> "#" <> pretty (review intIndex i)
-
-prettyCompTWithContext :: forall (ann :: Type). CompT Renamed -> PrettyM ann (Doc ann)
-prettyCompTWithContext (CompT count funArgs)
-  | review intCount count == 0 = prettyFunTy (NonEmpty.toNonEmpty funArgs)
-  | otherwise = bindVars count $ \newVars -> do
-      funTy <- prettyFunTy (NonEmpty.toNonEmpty funArgs) -- easier to pattern match
-      pure $ mkForall newVars funTy
-
-mkForall :: forall (ann :: Type). [Doc ann] -> Doc ann -> Doc ann
-mkForall tvars funTyBody = case tvars of
-  [] -> funTyBody
-  vars -> "forall" <+> hsep vars <> "." <+> funTyBody
-
-prettyFunTy :: forall (ann :: Type). NonEmpty (ValT Renamed) -> PrettyM ann (Doc ann)
-prettyFunTy (arg :| rest) = case rest of
-  [] -> ("!" <>) <$> prettyArg arg
-  (a : as) -> (\x y -> x <+> "->" <+> y) <$> prettyArg arg <*> prettyFunTy (a :| as)
-  where
-    prettyArg :: ValT Renamed -> PrettyM ann (Doc ann)
-    prettyArg vt
-      | isSimpleValT vt = prettyValTWithContext vt
-      | otherwise = parens <$> prettyValTWithContext vt
-
--- I.e. can we omit parens and get something unambiguous? This might be overly aggressive w/ parens but that's OK
-isSimpleValT :: forall (a :: Type). ValT a -> Bool
-isSimpleValT = \case
-  Abstraction _ -> True
-  BuiltinFlat _ -> True
-  ThunkT thunk -> isSimpleCompT thunk
-  where
-    isSimpleCompT :: CompT a -> Bool
-    isSimpleCompT (CompT count args) = review intCount count == 0 && NonEmpty.length args == 1
-
-prettyValTWithContext :: forall (ann :: Type). ValT Renamed -> PrettyM ann (Doc ann)
-prettyValTWithContext = \case
-  Abstraction abstr -> prettyRenamedWithContext abstr
-  ThunkT compT -> prettyCompTWithContext compT
-  BuiltinFlat biFlat -> pure $ viaShow biFlat
